+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Console Logger
+- Composite Logger
 
 ## [0.4.1] - 2024-10-26